Fix linting in init.c

+++ include/sys/syscall.h
@@ -223,8 +223,8 @@ void sched_yield();
 #endif //_SYSTEM
 
 #define __ALIGN1K(x) 	    (((((x))>>10)<<10)+1023)
-#define __get_env_address() (__ALIGN1K((int)get_sp()))
-#define __get_env_ptr()     (*((const char ***) __get_env_address()))
+#define __get_env_address() (__ALIGN1K((unsigned long)get_sp()))
+#define __get_env_ptr()     (*((const char ***) (unsigned long)__get_env_address()))
 #define __get_env()         (_environ ? _environ : __get_env_ptr()) 
 #define init_environ()      (_environ = __get_env())
